Adds scaffolding for part 1

+++ 2016-9_test.go
@@ -28,10 +28,10 @@ func Test2016_9_1(t *testing.T) {
 func Test2016_9_2(t *testing.T) {
 	var cases = []struct {
 		in   string
-		slen int
+		slen int64
 	}{
 		{"(3x3)XYZ", 9},
-		{"X(8x2)(3x3)ABCY", len("XABCABCABCABCABCABCY")},
+		{"X(8x2)(3x3)ABCY", int64(len("XABCABCABCABCABCABCY"))},
 		{"(27x12)(20x12)(13x14)(7x10)(1x12)A", 241920},
 		{"(25x3)(3x3)ABC(2x3)XY(5x2)PQRSTX(18x9)(3x2)TWO(5x7)SEVEN", 445},
 	}
